+++ src/utils_muon/__init__.py
@@ -1,3 +1,4 @@
-from .utils import adam_update, muon_update, hook_recorder, repeat_kv, override_model
+from .utils import adam_update, muon_update
+from .qk_hook import hook_recorder, repeat_kv, override_model
 
 __all__ = ["adam_update", "muon_update", "hook_recorder", "repeat_kv", "override_model"]
